update outputs, comments and added while loop

+++ pod-requeue.sh
@@ -6,37 +6,70 @@ else
   DRY_RUN=false
 fi
 
+POD_DUMP_RAW_JSON=pod-dump-raw.json
 POD_DUMP_JSON=pod-dump.json
 
+SLEEP=60
+
 export_pods() {
-    kubectl get po --export --all-namespaces -o json | jq '.items[] | select(.status.conditions[].reason) | select((.status.conditions[].reason == "Unschedulable") or (.status.conditions[].reason == "OutOfcpu")) |
-  del(
-    .status,
-    .spec.dnsPolicy,
-    .spec.securityContext,
-    .spec.terminationGracePeriodSeconds,
-    .spec.restartPolicy,
-    .metadata.uid,
-    .metadata.selfLink,
-    .metadata.resourceVersion,
-    .metadata.creationTimestamp,
-    .metadata.generation
+  # Collect list of all pods matching Status conditions reason of Unschedulable OR OutOfcpu
+    kubectl get po --export --all-namespaces -o json | jq '.items[] | select(.status.conditions[].reason) | select((.status.conditions[].reason == "Unschedulable") or (.status.conditions[].reason == "OutOfcpu"))' > $POD_DUMP_RAW_JSON
+}
+
+process_pods() {
+    # Remove server generated fields
+    cat $POD_DUMP_RAW_JSON | jq '
+    del(
+      .status,
+      .spec.dnsPolicy,
+      .spec.securityContext,
+      .spec.terminationGracePeriodSeconds,
+      .spec.restartPolicy,
+      .metadata.uid,
+      .metadata.selfLink,
+      .metadata.resourceVersion,
+      .metadata.creationTimestamp,
+      .metadata.generation
   )'  > $POD_DUMP_JSON
 }
 
-whack_pods() {
-	export_pods
-	POD_LIST=$(cat $POD_DUMP_JSON)
+pod_requeue() {
 
+  # Dry run loop
   if [[ $DRY_RUN == true ]]; then
-    echo "** Dry run: not executing. The following pods match for deletion:"
-		echo $POD_LIST | jq '(.metadata.name) | "Pod: \(.)"'
-		exit
+    while true; do
+      export_pods
+      RAW_POD_LIST=$(cat $POD_DUMP_RAW_JSON)
+      echo "** Dry run: not executing. The following pods match for deletion:"
+      cat $POD_DUMP_RAW_JSON | jq -r '[.metadata.name,.metadata.namespace,.status.conditions[].reason] | "Pod:\(.[0]) Namespace:\(.[1]) Reason:\(.[2])"'
+      echo "Sleeping for ${SLEEP} seconds"
+      sleep $SLEEP
+    done
+    exit
   fi
 
-  echo "Deleting stuck pods and recreating..."
-  kubectl delete -f $POD_DUMP_JSON && \
-  kubectl create -f $POD_DUMP_JSON
+  # Loop to find, delete and recreate matched pods
+  while true; do
+
+    # Collect matching pods and process
+    export_pods
+    process_pods
+
+    POD_LIST=$(cat $POD_DUMP_JSON)
+
+    # Do not run unless there's data in the processed pod dump file
+    if [ ! -s $POD_DUMP_JSON ]; then
+      echo "${POD_DUMP_JSON} is empty. Nothing to process"
+    else
+      echo "Deleting and recreating the following pods"
+      cat $POD_DUMP_RAW_JSON | jq -r '[.metadata.name,.metadata.namespace,.status.conditions[].reason] | "Pod:\(.[0]) Namespace:\(.[1]) Reason:\(.[2])"'
+      echo "---"
+      kubectl delete -f $POD_DUMP_JSON && \
+      kubectl create -f $POD_DUMP_JSON
+    fi
+    echo "---\nSleeping for ${SLEEP} seconds"
+    sleep $SLEEP
+  done
 }
 
-whack_pods $1
+pod_requeue $1
